fix analysis infinite spinner
