+++ applications/fishing-map/src/features/workspace/Workspace.tsx
@@ -9,33 +9,24 @@ import {
   selectWorkspaceStatus,
   selectWorkspaceError,
   selectWorkspace,
-  selectWorkspaceCustomStatus,
-  selectCurrentWorkspaceId,
 } from 'features/workspace/workspace.selectors'
 import { fetchResourceThunk } from 'features/resources/resources.slice'
 import { AsyncReducerStatus } from 'utils/async-slice'
-import { isGuestUser, isUserLogged } from 'features/user/user.selectors'
-import {
-  selectLocationCategory,
-  selectLocationType,
-  selectUrlViewport,
-  selectWorkspaceId,
-} from 'routes/routes.selectors'
-import { HOME, WORKSPACE } from 'routes/routes'
+import { isGuestUser } from 'features/user/user.selectors'
+import { selectLocationCategory, selectWorkspaceId } from 'routes/routes.selectors'
+import { HOME } from 'routes/routes'
 import { updateLocation } from 'routes/routes.actions'
 import { logoutUserThunk, selectUserData } from 'features/user/user.slice'
 import { selectSearchQuery } from 'features/app/app.selectors'
 import { SUPPORT_EMAIL } from 'data/config'
-import { DEFAULT_WORKSPACE_ID, WorkspaceCategories } from 'data/workspaces'
+import { WorkspaceCategories } from 'data/workspaces'
 import { selectDataviewsResourceQueries } from 'features/resources/resources.selectors'
-import useViewport from 'features/map/map-viewport.hooks'
 import HeatmapsSection from './heatmaps/HeatmapsSection'
 import VesselsSection from './vessels/VesselsSection'
 import EventsSection from './events/EventsSection'
 import EnvironmentalSection from './environmental/EnvironmentalSection'
 import ContextAreaSection from './context-areas/ContextAreaSection'
 import styles from './Workspace.module.css'
-import { fetchWorkspaceThunk } from './workspace.slice'
 
 function WorkspaceError(): React.ReactElement {
   const [logoutLoading, setLogoutLoading] = useState(false)
@@ -121,49 +112,11 @@ function WorkspaceError(): React.ReactElement {
 
 function Workspace() {
   const dispatch = useDispatch()
-  const { setMapCoordinates } = useViewport()
   const searchQuery = useSelector(selectSearchQuery)
-  const locationType = useSelector(selectLocationType)
   const workspace = useSelector(selectWorkspace)
   const workspaceStatus = useSelector(selectWorkspaceStatus)
   const locationCategory = useSelector(selectLocationCategory)
-  const currentWorkspaceId = useSelector(selectCurrentWorkspaceId)
-  const workspaceCustomStatus = useSelector(selectWorkspaceCustomStatus)
-  const userLogged = useSelector(isUserLogged)
-  const urlViewport = useSelector(selectUrlViewport)
-  const urlWorkspaceId = useSelector(selectWorkspaceId)
   const resourceQueries = useSelector(selectDataviewsResourceQueries)
-
-  const isHomeLocation = locationType === HOME
-  const homeNeedsFetch = isHomeLocation && currentWorkspaceId !== DEFAULT_WORKSPACE_ID
-  const hasWorkspaceIdChanged = locationType === WORKSPACE && currentWorkspaceId !== urlWorkspaceId
-  useEffect(() => {
-    let action: any
-    let actionResolved = false
-    const fetchWorkspace = async () => {
-      action = dispatch(fetchWorkspaceThunk(urlWorkspaceId as string))
-      const resolvedAction = await action
-      if (fetchWorkspaceThunk.fulfilled.match(resolvedAction)) {
-        if (!urlViewport && resolvedAction.payload?.viewport) {
-          setMapCoordinates(resolvedAction.payload.viewport)
-        }
-      }
-      actionResolved = true
-    }
-    if (
-      userLogged &&
-      workspaceCustomStatus !== AsyncReducerStatus.Loading &&
-      (homeNeedsFetch || hasWorkspaceIdChanged)
-    ) {
-      fetchWorkspace()
-    }
-    return () => {
-      if (action && action.abort !== undefined && !actionResolved) {
-        action.abort()
-      }
-    }
-    // eslint-disable-next-line react-hooks/exhaustive-deps
-  }, [userLogged, homeNeedsFetch, hasWorkspaceIdChanged])
 
   useEffect(() => {
     if (resourceQueries) {
